requirement.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-axes==7.0.2
django-environ==0.12.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
numpy==2.2.4
pandas==2.2.3
phonenumbers==9.0.3
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
django-storages==1.14.2
boto3==1.34.14
django-filter==25.1
python-magic==0.4.27
django-ckeditor==6.7.2
razorpay==1.4.2